Reorganize the query folder.
This changes the query folder to have a readers, writers and legacy
folder.

---
TYPE: NO_HISTORY
DESC: Reorganize the query folder.
