+++ test/src/unit-ReadCellSlabIter.cc
@@ -35,8 +35,8 @@
 #include "tiledb/common/common.h"
 #include "tiledb/sm/array_schema/tile_domain.h"
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
-#include "tiledb/sm/query/read_cell_slab_iter.h"
-#include "tiledb/sm/query/reader.h"
+#include "tiledb/sm/query/legacy/read_cell_slab_iter.h"
+#include "tiledb/sm/query/legacy/reader.h"
 
 #ifdef _WIN32
 #include "tiledb/sm/filesystem/win.h"
