+++ test/src/unit-capi-serialized_queries_using_subarray.cc
@@ -38,8 +38,8 @@
 #include "tiledb/sm/c_api/tiledb_serialization.h"
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
 #include "tiledb/sm/cpp_api/tiledb"
-#include "tiledb/sm/query/reader.h"
-#include "tiledb/sm/query/writer_base.h"
+#include "tiledb/sm/query/legacy/reader.h"
+#include "tiledb/sm/query/writers/writer_base.h"
 #include "tiledb/sm/serialization/query.h"
 
 #ifdef _WIN32
