+++ test/src/unit-result-coords.cc
@@ -33,8 +33,8 @@
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
 
 #include "test/src/helpers.h"
-#include "tiledb/sm/query/result_coords.h"
-#include "tiledb/sm/query/sparse_index_reader_base.h"
+#include "tiledb/sm/query/readers/result_coords.h"
+#include "tiledb/sm/query/readers/sparse_index_reader_base.h"
 
 #ifdef _WIN32
 #include "tiledb/sm/filesystem/win.h"
