+++ tiledb/sm/query/legacy/read_cell_slab_iter.cc
@@ -30,7 +30,7 @@
  * This file implements class ResultCellSlabIter.
  */
 
-#include "tiledb/sm/query/read_cell_slab_iter.h"
+#include "tiledb/sm/query/legacy/read_cell_slab_iter.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/sm/array/array.h"
 #include "tiledb/sm/array_schema/array_schema.h"
