+++ tiledb/sm/query/query.cc
@@ -40,14 +40,14 @@
 #include "tiledb/sm/enums/query_type.h"
 #include "tiledb/sm/fragment/fragment_metadata.h"
 #include "tiledb/sm/misc/parse_argument.h"
-#include "tiledb/sm/query/dense_reader.h"
-#include "tiledb/sm/query/global_order_writer.h"
-#include "tiledb/sm/query/ordered_writer.h"
+#include "tiledb/sm/query/legacy/reader.h"
 #include "tiledb/sm/query/query_condition.h"
-#include "tiledb/sm/query/reader.h"
-#include "tiledb/sm/query/sparse_global_order_reader.h"
-#include "tiledb/sm/query/sparse_unordered_with_dups_reader.h"
-#include "tiledb/sm/query/unordered_writer.h"
+#include "tiledb/sm/query/readers/dense_reader.h"
+#include "tiledb/sm/query/readers/sparse_global_order_reader.h"
+#include "tiledb/sm/query/readers/sparse_unordered_with_dups_reader.h"
+#include "tiledb/sm/query/writers/global_order_writer.h"
+#include "tiledb/sm/query/writers/ordered_writer.h"
+#include "tiledb/sm/query/writers/unordered_writer.h"
 #include "tiledb/sm/rest/rest_client.h"
 #include "tiledb/sm/storage_manager/storage_manager.h"
 #include "tiledb/sm/tile/writer_tile.h"
