+++ tiledb/sm/query/query_condition.h
@@ -41,8 +41,8 @@
 #include "tiledb/sm/array_schema/array_schema.h"
 #include "tiledb/sm/enums/query_condition_op.h"
 #include "tiledb/sm/query/ast/query_ast.h"
-#include "tiledb/sm/query/result_cell_slab.h"
-#include "tiledb/sm/query/result_tile.h"
+#include "tiledb/sm/query/readers/result_cell_slab.h"
+#include "tiledb/sm/query/readers/result_tile.h"
 
 using namespace tiledb::common;
