+++ tiledb/sm/query/readers/reader_base.h
@@ -34,7 +34,7 @@
 #define TILEDB_READER_BASE_H
 
 #include <queue>
-#include "strategy_base.h"
+#include "../strategy_base.h"
 #include "tiledb/common/common.h"
 #include "tiledb/common/status.h"
 #include "tiledb/sm/array_schema/dimension.h"
@@ -42,8 +42,8 @@
 #include "tiledb/sm/fragment/fragment_metadata.h"
 #include "tiledb/sm/misc/types.h"
 #include "tiledb/sm/query/query_condition.h"
-#include "tiledb/sm/query/result_cell_slab.h"
-#include "tiledb/sm/query/result_space_tile.h"
+#include "tiledb/sm/query/readers/result_cell_slab.h"
+#include "tiledb/sm/query/readers/result_space_tile.h"
 #include "tiledb/sm/subarray/subarray_partitioner.h"
 
 namespace tiledb {
