+++ tiledb/sm/query/readers/result_cell_slab.h
@@ -36,7 +36,7 @@
 #include <iostream>
 #include <vector>
 
-#include "tiledb/sm/query/result_tile.h"
+#include "tiledb/sm/query/readers/result_tile.h"
 
 using namespace tiledb::common;
