+++ tiledb/sm/query/readers/result_space_tile.h
@@ -40,7 +40,7 @@
 #include <vector>
 
 #include "tiledb/sm/misc/types.h"
-#include "tiledb/sm/query/result_tile.h"
+#include "tiledb/sm/query/readers/result_tile.h"
 
 using namespace tiledb::common;
