+++ tiledb/sm/query/readers/result_tile.cc
@@ -30,7 +30,7 @@
  * This file implements class ResultTile.
  */
 
-#include "tiledb/sm/query/result_tile.h"
+#include "tiledb/sm/query/readers/result_tile.h"
 #include "tiledb/sm/array_schema/dimension.h"
 #include "tiledb/sm/array_schema/domain.h"
 #include "tiledb/sm/enums/datatype.h"
