+++ tiledb/sm/query/readers/sparse_index_reader_base.cc
@@ -30,7 +30,7 @@
  * This file implements class SparseIndexReaderBase.
  */
 
-#include "tiledb/sm/query/sparse_index_reader_base.h"
+#include "tiledb/sm/query/readers/sparse_index_reader_base.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/common/memory_tracker.h"
 #include "tiledb/sm/array/array.h"
