+++ tiledb/sm/query/readers/sparse_unordered_with_dups_reader.cc
@@ -29,7 +29,7 @@
  *
  * This file implements class SparseUnorderedWithDupsReader.
  */
-#include "tiledb/sm/query/sparse_unordered_with_dups_reader.h"
+#include "tiledb/sm/query/readers/sparse_unordered_with_dups_reader.h"
 #include "tiledb/common/logger.h"
 #include "tiledb/common/memory_tracker.h"
 #include "tiledb/sm/array/array.h"
@@ -41,7 +41,7 @@
 #include "tiledb/sm/misc/parallel_functions.h"
 #include "tiledb/sm/misc/utils.h"
 #include "tiledb/sm/query/query_macros.h"
-#include "tiledb/sm/query/result_tile.h"
+#include "tiledb/sm/query/readers/result_tile.h"
 #include "tiledb/sm/stats/global_stats.h"
 #include "tiledb/sm/storage_manager/storage_manager.h"
 #include "tiledb/sm/subarray/subarray.h"
