+++ tiledb/sm/query/writers/global_order_writer.h
@@ -37,8 +37,8 @@
 
 #include "tiledb/common/common.h"
 #include "tiledb/common/status.h"
-#include "tiledb/sm/query/domain_buffer.h"
-#include "tiledb/sm/query/writer_base.h"
+#include "tiledb/sm/query/writers/domain_buffer.h"
+#include "tiledb/sm/query/writers/writer_base.h"
 
 using namespace tiledb::common;
