+++ tiledb/sm/query/writers/ordered_writer.cc
@@ -30,7 +30,7 @@
  * This file implements class OrderedWriter.
  */
 
-#include "tiledb/sm/query/ordered_writer.h"
+#include "tiledb/sm/query/writers/ordered_writer.h"
 #include "tiledb/common/common.h"
 #include "tiledb/common/heap_memory.h"
 #include "tiledb/common/logger.h"
