+++ tiledb/sm/query/writers/unordered_writer.cc
@@ -30,7 +30,7 @@
  * This file implements class UnorderedWriter.
  */
 
-#include "tiledb/sm/query/unordered_writer.h"
+#include "tiledb/sm/query/writers/unordered_writer.h"
 #include "tiledb/common/common.h"
 #include "tiledb/common/heap_memory.h"
 #include "tiledb/common/logger.h"
